% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coherence.R
\name{analyze_coherence_distribution}
\alias{analyze_coherence_distribution}
\title{Analyze Coherence Distribution}
\usage{
analyze_coherence_distribution(seurat_object, ecotype_column, sample_column)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{ecotype_column}{Ecotype column name}

\item{sample_column}{Sample column name}
}
\description{
Analyze Coherence Distribution
}
