% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{calculate_advanced_spatial_coherence}
\alias{calculate_advanced_spatial_coherence}
\title{Calculate Advanced Spatial Coherence}
\usage{
calculate_advanced_spatial_coherence(
  seurat_object,
  ecotype_column,
  sample_column,
  rand_num = 100,
  n_neighbors = 6,
  random_seed = NULL,
  verbose = TRUE
)

calculate_advanced_spatial_coherence(
  seurat_object,
  ecotype_column,
  sample_column,
  rand_num = 100,
  n_neighbors = 6,
  random_seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Seurat object with spatial data}

\item{ecotype_column}{Column name for ecotypes}

\item{sample_column}{Column name for samples}

\item{rand_num}{Number of random permutations}

\item{n_neighbors}{Number of spatial neighbors}

\item{random_seed}{Random seed for reproducibility}

\item{verbose}{Verbose output}
}
\description{
Calculate Advanced Spatial Coherence

Calculate Advanced Spatial Coherence
}
