% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{calculate_spatial_coherence}
\alias{calculate_spatial_coherence}
\title{Calculate Spatial Coherence}
\usage{
calculate_spatial_coherence(seurat_object, ecotype_column)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{ecotype_column}{Column name}
}
\description{
Calculate Spatial Coherence
}
