% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coherence.R
\name{compare_thresholds}
\alias{compare_thresholds}
\title{Compare Threshold Results}
\usage{
compare_thresholds(results_list, threshold_values)
}
\arguments{
\item{results_list}{List of analysis results}

\item{threshold_values}{Vector of threshold values}
}
\description{
Compare Threshold Results
}
