% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{create_compositional_differences}
\alias{create_compositional_differences}
\title{Create Compositional Differences Analysis}
\usage{
create_compositional_differences(
  seurat_object,
  coherence_results,
  ecotype_column,
  sample_column,
  coherence_thresholds = c(0.47, 0.39),
  output_dir = "spatial_coherence_results",
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{coherence_results}{Coherence results}

\item{ecotype_column}{Ecotype column name}

\item{sample_column}{Sample column name}

\item{coherence_thresholds}{Coherence thresholds for organization}

\item{output_dir}{Output directory}

\item{verbose}{Verbose output}
}
\description{
Create Compositional Differences Analysis
}
