% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{create_compositional_plot}
\alias{create_compositional_plot}
\title{Create Compositional Differences Plot}
\usage{
create_compositional_plot(differences_data)
}
\arguments{
\item{differences_data}{Compositional differences data}
}
\description{
Create Compositional Differences Plot
}
