% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{create_organization_heatmap}
\alias{create_organization_heatmap}
\title{Create Organization Heatmap}
\usage{
create_organization_heatmap(coherence_results)
}
\arguments{
\item{coherence_results}{Coherence results}
}
\description{
Create Organization Heatmap
}
