% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{create_publication_plots}
\alias{create_publication_plots}
\title{Create Publication Plots}
\usage{
create_publication_plots(
  coherence_results,
  organization_results,
  ecotype_annotations = NULL,
  treatment_results = NULL,
  compositional_results = NULL,
  sample_column = "sample"
)
}
\arguments{
\item{coherence_results}{Coherence results}

\item{organization_results}{Organization results}

\item{ecotype_annotations}{Ecotype annotations}

\item{treatment_results}{Treatment results}

\item{compositional_results}{Compositional analysis results}

\item{sample_column}{Sample column}
}
\description{
Create Publication Plots
}
