% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{get_package_version}
\alias{get_package_version}
\title{Get Package Version}
\usage{
get_package_version()

get_package_version()
}
\description{
Get Package Version

Get Package Version
}
