% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{neighbors_table_func_visium}
\alias{neighbors_table_func_visium}
\title{Calculate Distance-Based Neighbors for Visium Data}
\usage{
neighbors_table_func_visium(
  image_coords,
  spots_clusters,
  typical_distance = NULL,
  n_neighbors = 6
)

neighbors_table_func_visium(
  image_coords,
  spots_clusters,
  typical_distance = NULL,
  n_neighbors = 6
)
}
\arguments{
\item{image_coords}{Image coordinates dataframe}

\item{spots_clusters}{Spots clusters dataframe}

\item{typical_distance}{Typical distance between spots}

\item{n_neighbors}{Number of neighbors to consider}
}
\description{
Calculate Distance-Based Neighbors for Visium Data

Calculate Distance-Based Neighbors for Visium Data
}
