% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{obs_program_spatial_score}
\alias{obs_program_spatial_score}
\title{Calculate Observed Spatial Score}
\usage{
obs_program_spatial_score(program_neighbors, cluster)

obs_program_spatial_score(program_neighbors, cluster)
}
\arguments{
\item{program_neighbors}{Program neighbors}

\item{cluster}{Cluster}
}
\description{
Calculate Observed Spatial Score

Calculate Observed Spatial Score
}
