% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{one_val}
\alias{one_val}
\title{Calculate Theoretical Maximum (one_val)}
\usage{
one_val(spots_num)

one_val(spots_num)
}
\arguments{
\item{spots_num}{Number of spots}
}
\description{
Calculate Theoretical Maximum (one_val)

Calculate Theoretical Maximum (one_val)
}
