% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coherence.R
\name{perform_threshold_analysis}
\alias{perform_threshold_analysis}
\title{Perform Threshold Analysis}
\usage{
perform_threshold_analysis(
  seurat_object,
  ecotype_column,
  sample_column,
  thresholds = c(0.3, 0.35, 0.4, 0.45, 0.5, 0.55, 0.6)
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{ecotype_column}{Ecotype column name}

\item{sample_column}{Sample column name}

\item{thresholds}{Vector of thresholds to test}
}
\description{
Perform Threshold Analysis
}
