% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{plot_organization_spectrum}
\alias{plot_organization_spectrum}
\title{Plot Organization Spectrum}
\usage{
plot_organization_spectrum(coherence_scores, organization_classes)
}
\arguments{
\item{coherence_scores}{Scores}

\item{organization_classes}{Classes}
}
\description{
Plot Organization Spectrum
}
