% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{run_spatial_analysis}
\alias{run_spatial_analysis}
\title{Run Comprehensive Spatial Analysis}
\usage{
run_spatial_analysis(
  seurat_object,
  ecotype_column,
  sample_column = "orig.ident",
  treatment_column = NULL,
  output_dir = "spatial_coherence_results",
  save_csvs = TRUE,
  create_plots = TRUE,
  config_path = NULL,
  use_real_algorithm = TRUE,
  n_permutations = 100,
  parallel = FALSE,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Seurat object with spatial data}

\item{ecotype_column}{Column name for ecotypes}

\item{sample_column}{Column name for samples}

\item{treatment_column}{Column name for treatment (optional)}

\item{output_dir}{Directory to save results and plots}

\item{save_csvs}{Whether to save CSV files (default: TRUE)}

\item{create_plots}{Whether to create plots (default: TRUE)}

\item{config_path}{Config path (optional)}

\item{use_real_algorithm}{Use real algorithm}

\item{n_permutations}{Number of permutations}

\item{parallel}{Use parallel}

\item{n_cores}{Number of cores}

\item{verbose}{Verbose}
}
\description{
Run Comprehensive Spatial Analysis
}
