% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{sample_programs_composition}
\alias{sample_programs_composition}
\title{Calculate Programs Composition}
\usage{
sample_programs_composition(spots_clusters, gen_clusters)

sample_programs_composition(spots_clusters, gen_clusters)
}
\arguments{
\item{spots_clusters}{Spots clusters dataframe}

\item{gen_clusters}{All possible clusters}
}
\description{
Calculate Programs Composition

Calculate Programs Composition
}
