% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coherence.R
\name{save_publication_plots}
\alias{save_publication_plots}
\title{Save Publication Plots}
\usage{
save_publication_plots(
  results,
  base_name = "SpatialCoherence",
  output_dir = ".",
  formats = c("pdf", "png"),
  width = 12,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{results}{Analysis results object}

\item{base_name}{Base filename}

\item{output_dir}{Output directory}

\item{formats}{Vector of formats to save}

\item{width}{Plot width}

\item{height}{Plot height}

\item{dpi}{Resolution}
}
\description{
Save Publication Plots
}
