% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{validate_spatial_data}
\alias{validate_spatial_data}
\title{Validate Spatial Data}
\usage{
validate_spatial_data(seurat_object, min_spots = 10, min_genes = 100)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{min_spots}{Minimum spots}

\item{min_genes}{Minimum genes}
}
\description{
Validate Spatial Data
}
