% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{zero_val}
\alias{zero_val}
\title{Calculate Expected Random Value (zero_val)}
\usage{
zero_val(rand_table, spots_clusters, cluster)

zero_val(rand_table, spots_clusters, cluster)
}
\arguments{
\item{rand_table}{Random table}

\item{spots_clusters}{Spots clusters}

\item{cluster}{Cluster}
}
\description{
Calculate Expected Random Value (zero_val)

Calculate Expected Random Value (zero_val)
}
