% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{classify_organization}
\alias{classify_organization}
\title{Classify Organization}
\usage{
classify_organization(
  coherence_scores,
  method = "threshold",
  threshold = 0.47,
  abundance_threshold = 0.05
)
}
\arguments{
\item{coherence_scores}{Coherence scores}

\item{method}{Method}

\item{threshold}{Threshold}

\item{abundance_threshold}{Abundance threshold}
}
\description{
Classify Organization
}
