% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R
\name{run_spatial_analysis}
\alias{run_spatial_analysis}
\title{Run Spatial Analysis}
\usage{
run_spatial_analysis(
  seurat_object,
  config_path = NULL,
  output_dir = NULL,
  use_real_algorithm = TRUE,
  n_permutations = 100,
  parallel = FALSE,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{config_path}{Config path}

\item{output_dir}{Output directory}

\item{use_real_algorithm}{Use real algorithm}

\item{n_permutations}{Number of permutations}

\item{parallel}{Use parallel}

\item{n_cores}{Number of cores}

\item{verbose}{Verbose}
}
\description{
Run Spatial Analysis
}
