% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCoherence.R, R/spatial_coherence.R
\name{validate_spatial_data}
\alias{validate_spatial_data}
\title{Validate Spatial Data}
\usage{
validate_spatial_data(
  seurat_object,
  min_spots = 50,
  min_genes = 100,
  ecotype_column = NULL,
  sample_column = NULL
)

validate_spatial_data(
  seurat_object,
  min_spots = 50,
  min_genes = 100,
  ecotype_column = NULL,
  sample_column = NULL
)
}
\arguments{
\item{seurat_object}{Seurat object}

\item{min_spots}{Minimum spots per ecotype}

\item{min_genes}{Minimum genes (legacy parameter)}

\item{ecotype_column}{Ecotype column name to check}

\item{sample_column}{Sample column name to check}
}
\description{
Validate Spatial Data

Enhanced Validate Spatial Data
}
